added material class for each material type
